+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New method `Redmine\Api\IssuePriority::fromHttpClient()` for creating the class.
 - New method `Redmine\Api\IssueRelation::fromHttpClient()` for creating the class.
 - New method `Redmine\Api\IssueStatus::fromHttpClient()` for creating the class.
+- New method `Redmine\Api\Membership::fromHttpClient()` for creating the class.
 - Add support for PHP 8.5
 - Add support for Redmine 6.1.
 
@@ -42,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Extending `Redmine\Api\IssueRelation` is deprecated and will be set to final in future, create a wrapper class instead.
 - `Redmine\Api\IssueStatus::__construct()` is deprecated and will be set to private in future, use `\Redmine\Api\IssueStatus::fromHttpClient()` instead.
 - Extending `Redmine\Api\IssueStatus` is deprecated and will be set to final in future, create a wrapper class instead.
+- `Redmine\Api\Membership::__construct()` is deprecated and will be set to private in future, use `\Redmine\Api\Membership::fromHttpClient()` instead.
+- Extending `Redmine\Api\Membership` is deprecated and will be set to final in future, create a wrapper class instead.
 
 ### Removed